chore: move clear routing data to debug tools
