+++ frontend/src/screens/channels/Channels.tsx
@@ -19,17 +19,6 @@ import EmptyState from "src/components/EmptyState.tsx";
 import ExternalLink from "src/components/ExternalLink";
 import Loading from "src/components/Loading.tsx";
 import {
-  AlertDialog,
-  AlertDialogAction,
-  AlertDialogCancel,
-  AlertDialogContent,
-  AlertDialogDescription,
-  AlertDialogFooter,
-  AlertDialogHeader,
-  AlertDialogTitle,
-  AlertDialogTrigger,
-} from "src/components/ui/alert-dialog";
-import {
   Alert,
   AlertDescription,
   AlertTitle,
@@ -51,16 +40,8 @@ import {
   DropdownMenuSeparator,
   DropdownMenuTrigger,
 } from "src/components/ui/dropdown-menu.tsx";
-import { Label } from "src/components/ui/label";
 import { LoadingButton } from "src/components/ui/loading-button.tsx";
 import { CircleProgress } from "src/components/ui/progress.tsx";
-import {
-  Select,
-  SelectContent,
-  SelectItem,
-  SelectTrigger,
-  SelectValue,
-} from "src/components/ui/select";
 import {
   Tooltip,
   TooltipContent,
@@ -75,7 +56,6 @@ import {
 import { useAlbyBalance } from "src/hooks/useAlbyBalance.ts";
 import { useBalances } from "src/hooks/useBalances.ts";
 import { useChannels } from "src/hooks/useChannels";
-import { useInfo } from "src/hooks/useInfo";
 import { useIsDesktop } from "src/hooks/useMediaQuery.ts";
 import { useNodeConnectionInfo } from "src/hooks/useNodeConnectionInfo.ts";
 import { useRedeemOnchainFunds } from "src/hooks/useRedeemOnchainFunds.ts";
@@ -85,25 +65,6 @@ import { cn } from "src/lib/utils.ts";
 import { Channel, Node, UpdateChannelRequest } from "src/types";
 import { request } from "src/utils/request";
 
-const RESET_KEY_OPTIONS = [
-  {
-    value: "ALL",
-    label: "All",
-  },
-  {
-    value: "LatestRgsSyncTimestamp",
-    label: "LatestRgsSyncTimestamp",
-  },
-  {
-    value: "Scorer",
-    label: "Scorer",
-  },
-  {
-    value: "NetworkGraph",
-    label: "NetworkGraph",
-  },
-];
-
 export default function Channels() {
   useSyncWallet();
   const { data: channels, mutate: reloadChannels } = useChannels();
@@ -111,13 +72,11 @@ export default function Channels() {
   const { data: balances } = useBalances();
   const { data: albyBalance, mutate: reloadAlbyBalance } = useAlbyBalance();
   const [nodes, setNodes] = React.useState<Node[]>([]);
-  const { mutate: reloadInfo } = useInfo();
 
   const redeemOnchainFunds = useRedeemOnchainFunds();
   const { toast } = useToast();
   const [drainingAlbySharedFunds, setDrainingAlbySharedFunds] =
     React.useState(false);
-  const [resetKey, setResetKey] = React.useState<string>();
   const isDesktop = useIsDesktop();
 
   const nodeHealth = channels ? getNodeHealth(channels) : 0;
@@ -187,26 +146,6 @@ export default function Channels() {
     }
   }
 
-  async function resetRouter() {
-    try {
-      await request("/api/reset-router", {
-        method: "POST",
-        body: JSON.stringify({ key: resetKey }),
-        headers: {
-          "Content-Type": "application/json",
-        },
-      });
-      await reloadInfo();
-      toast({ title: "🎉 Router reset" });
-    } catch (error) {
-      console.error(error);
-      toast({
-        variant: "destructive",
-        title: "Something went wrong: " + error,
-      });
-    }
-  }
-
   const showHostedBalance =
     albyBalance && albyBalance.sats > ALBY_HIDE_HOSTED_BALANCE_LIMIT;
 
@@ -217,123 +156,72 @@ export default function Channels() {
         description="Manage your lightning node"
         contentRight={
           <div className="flex gap-3 items-center justify-center">
-            <AlertDialog>
-              <DropdownMenu>
-                <DropdownMenuTrigger asChild>
-                  <Button variant="outline" size="default">
-                    Advanced
-                    <ChevronDown />
-                  </Button>
-                </DropdownMenuTrigger>
-                <DropdownMenuContent className="w-56">
-                  <DropdownMenuGroup>
-                    <DropdownMenuItem>
-                      <div
-                        className="flex flex-row gap-4 items-center w-full cursor-pointer"
-                        onClick={() => {
-                          if (!nodeConnectionInfo) {
-                            return;
-                          }
-                          copyToClipboard(nodeConnectionInfo.pubkey);
-                        }}
-                      >
-                        <div>Node</div>
-                        <div className="overflow-hidden text-ellipsis flex-1">
-                          {nodeConnectionInfo?.pubkey || "Loading..."}
-                        </div>
-                        {nodeConnectionInfo && (
-                          <CopyIcon className="shrink-0 w-4 h-4" />
-                        )}
+            <DropdownMenu>
+              <DropdownMenuTrigger asChild>
+                <Button variant="outline" size="default">
+                  Advanced
+                  <ChevronDown />
+                </Button>
+              </DropdownMenuTrigger>
+              <DropdownMenuContent className="w-56">
+                <DropdownMenuGroup>
+                  <DropdownMenuItem>
+                    <div
+                      className="flex flex-row gap-4 items-center w-full cursor-pointer"
+                      onClick={() => {
+                        if (!nodeConnectionInfo) {
+                          return;
+                        }
+                        copyToClipboard(nodeConnectionInfo.pubkey);
+                      }}
+                    >
+                      <div>Node</div>
+                      <div className="overflow-hidden text-ellipsis flex-1">
+                        {nodeConnectionInfo?.pubkey || "Loading..."}
                       </div>
-                    </DropdownMenuItem>
-                  </DropdownMenuGroup>
-                  <DropdownMenuGroup>
-                    <DropdownMenuSeparator />
-                    <DropdownMenuItem>
-                      <Link
-                        to="/channels/onchain/deposit-bitcoin"
-                        className="w-full"
-                      >
-                        Deposit Bitcoin
-                      </Link>
-                    </DropdownMenuItem>
-                    {(balances?.onchain.spendable || 0) > ONCHAIN_DUST_SATS && (
-                      <DropdownMenuItem
-                        onClick={redeemOnchainFunds.redeemFunds}
-                        disabled={redeemOnchainFunds.isLoading}
-                        className="w-full cursor-pointer"
-                      >
-                        Withdraw Savings Balance
-                        {redeemOnchainFunds.isLoading && <Loading />}
-                      </DropdownMenuItem>
-                    )}
-                  </DropdownMenuGroup>
-                  <DropdownMenuSeparator />
-                  <DropdownMenuGroup>
-                    <DropdownMenuLabel>Management</DropdownMenuLabel>
-                    <DropdownMenuItem>
-                      <Link className="w-full" to="/peers">
-                        Connected Peers
-                      </Link>
-                    </DropdownMenuItem>
-                    <DropdownMenuItem>
-                      <Link className="w-full" to="/wallet/sign-message">
-                        Sign Message
-                      </Link>
-                    </DropdownMenuItem>
-                    <AlertDialogTrigger asChild>
-                      <DropdownMenuItem className="w-full cursor-pointer">
-                        Clear Routing Data
-                      </DropdownMenuItem>
-                    </AlertDialogTrigger>
-                  </DropdownMenuGroup>
-                </DropdownMenuContent>
-              </DropdownMenu>
-              <AlertDialogContent>
-                <AlertDialogHeader>
-                  <AlertDialogTitle>Clear Routing Data</AlertDialogTitle>
-                  <AlertDialogDescription>
-                    <div>
-                      <p>Are you sure you want to clear your routing data?</p>
-                      <div className="grid gap-2 mt-4">
-                        <Label className="text-primary">
-                          Routing Data to Clear
-                        </Label>
-                        <Select
-                          name="resetKey"
-                          value={resetKey}
-                          onValueChange={(value) => setResetKey(value)}
-                        >
-                          <SelectTrigger>
-                            <SelectValue placeholder="Select Data" />
-                          </SelectTrigger>
-                          <SelectContent>
-                            {RESET_KEY_OPTIONS.map((resetKey) => (
-                              <SelectItem
-                                key={resetKey.value}
-                                value={resetKey.value}
-                              >
-                                {resetKey.label}
-                              </SelectItem>
-                            ))}
-                          </SelectContent>
-                        </Select>
-                      </div>
-                      <p className="text-primary font-medium mt-4">
-                        After clearing, you'll need to login again to restart
-                        your node.
-                      </p>
+                      {nodeConnectionInfo && (
+                        <CopyIcon className="shrink-0 w-4 h-4" />
+                      )}
                     </div>
-                  </AlertDialogDescription>
-                </AlertDialogHeader>
-                <AlertDialogFooter>
-                  <AlertDialogCancel>Cancel</AlertDialogCancel>
-                  <AlertDialogAction disabled={!resetKey} onClick={resetRouter}>
-                    Confirm
-                  </AlertDialogAction>
-                </AlertDialogFooter>
-              </AlertDialogContent>
-            </AlertDialog>
+                  </DropdownMenuItem>
+                </DropdownMenuGroup>
+                <DropdownMenuGroup>
+                  <DropdownMenuSeparator />
+                  <DropdownMenuItem>
+                    <Link
+                      to="/channels/onchain/deposit-bitcoin"
+                      className="w-full"
+                    >
+                      Deposit Bitcoin
+                    </Link>
+                  </DropdownMenuItem>
+                  {(balances?.onchain.spendable || 0) > ONCHAIN_DUST_SATS && (
+                    <DropdownMenuItem
+                      onClick={redeemOnchainFunds.redeemFunds}
+                      disabled={redeemOnchainFunds.isLoading}
+                      className="w-full cursor-pointer"
+                    >
+                      Withdraw Savings Balance
+                      {redeemOnchainFunds.isLoading && <Loading />}
+                    </DropdownMenuItem>
+                  )}
+                </DropdownMenuGroup>
+                <DropdownMenuSeparator />
+                <DropdownMenuGroup>
+                  <DropdownMenuLabel>Management</DropdownMenuLabel>
+                  <DropdownMenuItem>
+                    <Link className="w-full" to="/peers">
+                      Connected Peers
+                    </Link>
+                  </DropdownMenuItem>
+                  <DropdownMenuItem>
+                    <Link className="w-full" to="/wallet/sign-message">
+                      Sign Message
+                    </Link>
+                  </DropdownMenuItem>
+                </DropdownMenuGroup>
+              </DropdownMenuContent>
+            </DropdownMenu>
             <Link to="/channels/outgoing">
               <Button>Open Channel</Button>
             </Link>
